move webpack-related fns to utils/webpack
